chore: Refactor tournament system for improved readability and maintainability
